+++ src/base/objective/first_order_nonsmooth.jl
@@ -1,17 +1,6 @@
-# TODO: subgrad, prox
-
-function _ensure_mutating_prox(prox_f, p, evaluation::AbstractEvaluationType)
-    return prox_f
-end
-function _ensure_mutating_prox(prox_f, q::Number, evaluation::AllocatingEvaluation)
-    return isnothing(prox_f) ? prox_f : (M, λ, p) -> [prox_f(M, λ, p[])]
-end
-function _ensure_mutating_prox(prox_f, q::Number, evaluation::InplaceEvaluation)
-    return isnothing(prox_f) ? prox_f : (M, q, λ, p) -> (q .= [prox_f(M, λ, p[])])
-end
-
 function get_proximal_map end
-#TODO: Add a docstring also here?
+
+#TODO: Add a docstring aløso here
 function get_proximal_map(amp::AbstractManoptProblem, λ, p, i)
     return get_proximal_map(get_manifold(amp), get_objective(amp), λ, p, i)
 end
@@ -57,8 +46,10 @@ function get_subgradient!(
     )
     return get_subgradient!(M, X, get_objective(admo, false), p)
 end
-function get_subgradient_function(admo::AbstractDecoratedManifoldObjective, recursive = false)
-    return get_subgradient_function(get_objective(admo, recursive))
+function get_subgradient_function(
+        admo::AbstractDecoratedManifoldObjective, recursive = false; evaluation::AbstractEvaluationType = AllocatingEvaluation()
+    )
+    return get_subgradient_function(get_objective(admo, recursive); evaluation = evaluation)
 end
 
 #